1.refactor: 优化项目整体结构，删除无用组件 2.refactor: 优化交互细节
