--- src/views/admin/Article.vue
@@ -1,31 +1,0 @@
-<template>
-  <div class="container">
-    <!-- 面包屑导航区域 -->
-    <el-breadcrumb separator-class="el-icon-arrow-right">
-      <el-breadcrumb-item :to="{ path: '/' }">首页</el-breadcrumb-item>
-      <el-breadcrumb-item>文章管理</el-breadcrumb-item>
-      <el-breadcrumb-item>文章列表</el-breadcrumb-item>
-    </el-breadcrumb>
-    <!-- 主体部分 -->
-    <Content type="article"></Content>
-  </div>
-</template>
-
-<script>
-import Content from '@/components/admin/common/Content.vue'
-export default {
-  components: {
-    Content
-  }
-}
-</script>
-
-<style lang="less" scoped>
-.container {
-  padding: 20px;
-}
-.el-breadcrumb {
-  margin-bottom: 15px;
-  font-size: 15px;
-}
-</style>
